made to do list into seperate components for mobile and desktop
